```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkFieldSet/serialize-20-12         15649         15722         +0.47%
BenchmarkFieldSet/deserialize-20-12       36035         26974         -25.14%
BenchmarkFieldSet/serialize-50-12         37865         38637         +2.04%
BenchmarkFieldSet/deserialize-50-12       81147         66082         -18.57%
BenchmarkFieldSet/serialize-100-12        114218        118479        +3.73%
BenchmarkFieldSet/deserialize-100-12      224422        192446        -14.25%
BenchmarkFieldSet/serialize-500-12        617479        611650        -0.94%
BenchmarkFieldSet/deserialize-500-12      1053144       842299        -20.02%
BenchmarkFieldSet/serialize-1000-12       1236755       1160060       -6.20%
BenchmarkFieldSet/deserialize-1000-12     2054208       1669777       -18.71%

benchmark                                 old allocs     new allocs     delta
BenchmarkFieldSet/serialize-20-12         23             23             +0.00%
BenchmarkFieldSet/deserialize-20-12       216            141            -34.72%
BenchmarkFieldSet/serialize-50-12         48             48             +0.00%
BenchmarkFieldSet/deserialize-50-12       614            405            -34.04%
BenchmarkFieldSet/serialize-100-12        93             94             +1.08%
BenchmarkFieldSet/deserialize-100-12      2166           1441           -33.47%
BenchmarkFieldSet/serialize-500-12        470            470            +0.00%
BenchmarkFieldSet/deserialize-500-12      10779          7137           -33.79%
BenchmarkFieldSet/serialize-1000-12       999            998            -0.10%
BenchmarkFieldSet/deserialize-1000-12     23523          15604          -33.66%
```

This depends on https://github.com/json-iterator/go/pull/394. It doesn't really move the needle on the main repo benchmarks, sadly. But I already did the work, so I'm just publishing it.