@lavalamp 2019-07-26
Actually let me retract that, it *does* do some good in the main repo, I had screwed up my go.mod rewrite rule:

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1953454       1928973       -1.25%
BenchmarkUpdateNewObject-12     2242290       2161230       -3.62%
BenchmarkRepeatedUpdate-12      781621        747257        -4.40%
BenchmarkSetToFields-12         5136          5359          +4.34%
BenchmarkFieldsToSet-12         10061         6531          -35.09%
BenchmarkYAMLToTyped-12         61988         63139         +1.86%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      6827           6827           +0.00%
BenchmarkUpdateNewObject-12     7913           7892           -0.27%
BenchmarkRepeatedUpdate-12      2288           2257           -1.35%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             59             -28.05%
BenchmarkYAMLToTyped-12         152            152            +0.00%
```