@lavalamp 2019-07-26
The aggregate, overall improvement is:

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      2230683       1928973       -13.53%
BenchmarkUpdateNewObject-12     2624417       2161230       -17.65%
BenchmarkRepeatedUpdate-12      1031653       747257        -27.57%
BenchmarkSetToFields-12         24778         5359          -78.37%
BenchmarkFieldsToSet-12         33876         6531          -80.72%
BenchmarkYAMLToTyped-12         68377         63139         -7.66%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      8632           6827           -20.91%
BenchmarkUpdateNewObject-12     11210          7892           -29.60%
BenchmarkRepeatedUpdate-12      3057           2257           -26.17%
BenchmarkSetToFields-12         124            14             -88.71%
BenchmarkFieldsToSet-12         147            59             -59.86%
BenchmarkYAMLToTyped-12         150            152            +1.33%
```